requirements.txt
# Core Framework
fastapi
uvicorn[standard]

# Database
SQLAlchemy
alembic
psycopg2-binary

# Environment Variables
python-dotenv

# File Uploads
python-multipart

# Authentication & Security
passlib[bcrypt]
passlib[argon2]  
python-jose[cryptography]
bcrypt==4.1.2
passlib==1.7.4

# Image / AR Processing
Pillow

# Development & Debugging
httpx
pydantic
pydantic-settings
email-validator
request